requirements.txt
AuthorizeSauce==0.4.1
BeautifulSoup==3.2.1
cssselect==0.9.1
Django==1.8.6
MySQL-python==1.2.5
PyYAML==3.11
South==1.0.2
beautifulsoup4==4.4.1
bleach==1.4.2
boto==2.38.0
celery==3.1.19
coverage==4.0.2
defusedxml==0.4.1
django-celery==3.1.17
django-debug-toolbar==1.4
django-extensions==1.5.9
django-haystack==2.4.1
django-jenkins==0.18.1
django-passwords==0.3.4
django-memcached==0.1.2
django-nose==1.4.2
django-recaptcha==1.0.4
django-storages==1.1.8
django-taggit==0.17.3
django-tastypie==0.12.2
django-widget-tweaks==1.4.1
django-cors-headers==1.1.0
factory_boy==2.6.0
git+git://github.com/DirectEmployers/django-fsm-widget.git@dc928b0b034a3f4c456e83b3e4511c53a24b98a5
git+git://github.com/DirectEmployers/saved-search.git
git+git://github.com/DirectEmployers/solrsitemap.git
git+git://github.com/DirectEmployers/unicode-slugify.git
git+git://github.com/arneb/django-messages.git@bc443623fde4945a841c48665b477a529f4ec676
git+git://github.com/frankban/django-endless-pagination.git@725bde91db6da2be0663652b0df4501ba72306e7
jira==1.0.3
lxml==3.5.0
markdown2==2.3.0
mock==1.3.0
newrelic==2.58.0.43
ordereddict==1.1
pep8==1.6.2
pydns==2.3.6
pyflakes==1.0.0
pynliner==0.5.2
pysolr==3.3.3
python-memcached==1.57
requests==2.8.1
selenium==2.48.0
unicodecsv==0.14.1
bidict==0.3.1
validate-email==1.3
testfixtures==4.5.0
django-compressor==1.5
Markdown==2.6.4
xlrd==0.9.4
